+++ shaketune/graph_creators/graph_static.py
