+++ shaketune/shaketune.py
@@ -4,21 +4,21 @@
 import os
 from pathlib import Path
 
-from .helpers.console_output import ConsoleOutput
-from .measurement import (
+from .commands import (
     axes_map_calibration,
     axes_shaper_calibration,
     compare_belts_responses,
     create_vibrations_profile,
     excitate_axis_at_freq,
 )
-from .post_processing import (
+from .graph_creators import (
     AxesMapFinder,
     BeltsGraphCreator,
     ShaperGraphCreator,
     StaticGraphCreator,
     VibrationsGraphCreator,
 )
+from .helpers.console_output import ConsoleOutput
 from .shaketune_config import ShakeTuneConfig
 from .shaketune_process import ShakeTuneProcess
